Created console and WPF GUI app
Basically you can choose a DLL and methods from it, and then provide
values for the parameters and run it.

Both are really proof of concepts at this point and I'm sure have many
many bugs.  There is no error checking/validation really at this point.
And it is ugly, oh very ugly.